% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uz2_circ.R
\name{uz2_circ}
\alias{uz2_circ}
\title{Bayesian inference for the Circular Type 2 moult model}
\usage{
uz2_circ(
  moult_index_column,
  date_column,
  start_formula = ~1,
  duration_formula = ~1,
  kappa_formula = ~1,
  duration_prior = list(mean = 182.5, sd = 61),
  lump_non_moult = FALSE,
  data,
  init = "auto",
  log_lik = TRUE,
  opt = TRUE,
  ...
)
}
\arguments{
\item{moult_index_column}{the name the column in `data` containing moult indices, i.e. a numeric vector of (linearized) moult scores (0 = old plumage,1 = new plumage).}

\item{date_column}{the name the column in `data` containing sampling dates, encoded as days since an arbitrary reference date, i.e. a numeric vector}

\item{start_formula}{model formula for start date}

\item{duration_formula}{model formula for duration}

\item{kappa_formula}{model formula for start date sigma}

\item{duration_prior}{list with elements `mean` and `sd`. Hyperparameters for the normal prior on duration (in days; 0 < mean < 365; sd > 0)}

\item{lump_non_moult}{logical; should pre- and post-moult observations be treated as indistinguishable? if TRUE, the type 2L model will be fitted.}

\item{data}{Input data frame must contain a numeric column "date" and a column "moult_index" which is a numeric vector of moult scores ranging from 0 (old plumage) to  1 (new plumage).}

\item{init}{Specification of initial values for all or some parameters. Can be the string "auto" for an automatic guess based on the data, or any of the permitted rstan options: the digit 0, the strings "0" or "random", or a function. See the detailed documentation for the init argument in ?rstan::stan.}

\item{log_lik}{boolean retain pointwise log-likelihood in output? This enables model assessment and selection via the loo package. Defaults to true, can lead to very large output arrays if sample size is large.}

\item{opt}{logical use "optimized version of stancode}

\item{...}{Arguments passed to `cmdstanr` sampling method (e.g. chains).}
}
\value{
An object of class `moultmcmc`
}
\description{
Bayesian inference for the Circular Type 2 moult model
}
